@@ -6,40 +6,42 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/gruntwork-io/terragrunt/codegen"
+	"github.com/gruntwork-io/terragrunt/config/hclparse"
+	"github.com/gruntwork-io/terragrunt/pkg/log"
+
+	"dario.cat/mergo"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
-	"github.com/imdario/mergo"
-	"github.com/zclconf/go-cty/cty"
 
-	"github.com/gruntwork-io/terragrunt/errors"
+	"maps"
+
+	"github.com/gruntwork-io/terragrunt/internal/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/util"
 )
 
 const bareIncludeKey = ""
 
+var fieldsCopyLocks = util.NewKeyLocks()
+
 // Parse the config of the given include, if one is specified
-func parseIncludedConfig(
-	includedConfig *IncludeConfig,
-	terragruntOptions *options.TerragruntOptions,
-	dependencyOutputs *cty.Value,
-	decodeList []PartialDecodeSectionType,
-) (*TerragruntConfig, error) {
+func parseIncludedConfig(ctx *ParsingContext, l log.Logger, includedConfig *IncludeConfig) (*TerragruntConfig, error) {
 	if includedConfig.Path == "" {
-		return nil, errors.WithStackTrace(IncludedConfigMissingPath(terragruntOptions.TerragruntConfigPath))
+		return nil, errors.New(IncludedConfigMissingPathError(ctx.TerragruntOptions.TerragruntConfigPath))
 	}
 
 	includePath := includedConfig.Path
 
 	if !filepath.IsAbs(includePath) {
-		includePath = util.JoinPath(filepath.Dir(terragruntOptions.TerragruntConfigPath), includePath)
+		includePath = util.JoinPath(filepath.Dir(ctx.TerragruntOptions.TerragruntConfigPath), includePath)
 	}
 
-	// These condition are here to specifically handle the `run-all` command. During any `run-all` call, terragrunt
+	// These condition are here to specifically handle the `run --all` command. During any `run --all` call, terragrunt
 	// needs to first build up the dependency graph to know what order to process the modules in. We want to limit users
 	// from creating a dependency between the dependency path for graph generation, and a module output. This is because
 	// the outputs may not be available yet during the graph generation. E.g., consider a completely new deployment and
-	// `terragrunt run-all apply` is called. In this case, the outputs are expected to be materialized while terragrunt
+	// `terragrunt run --all apply` is called. In this case, the outputs are expected to be materialized while terragrunt
 	// is running `apply` through the graph, but NOT when the dependency graph is first being formulated.
 	//
 	// To support this, we implement the following conditions for when terragrunt can fully parse the included config
@@ -53,7 +55,7 @@ func parseIncludedConfig(
 	// cause this! For example, suppose the user has a dependency path that depends on an included input:
 	//
 	// include "root" {
-	//   path = find_in_parent_folders()
+	//   path = find_in_parent_folders("root.hcl")
 	//   expose = true
 	// }
 	// dependency "dep" {
@@ -79,113 +81,99 @@ func parseIncludedConfig(
 	// NOTE: To make the logic easier to implement, we implement the inverse here, where we check whether the included
 	// config has a dependency block, and if we are in the middle of a partial parse, we perform a partial parse of the
 	// included config.
-	hasDependency, err := configFileHasDependencyBlock(includePath, terragruntOptions)
+	hasDependency, err := configFileHasDependencyBlock(includePath)
 	if err != nil {
 		return nil, err
 	}
-	if hasDependency && len(decodeList) > 0 {
-		terragruntOptions.Logger.Debugf(
-			"Included config %s can only be partially parsed during dependency graph formation for run-all command as it has a dependency block.",
+
+	if hasDependency && len(ctx.PartialParseDecodeList) > 0 {
+		l.Debugf(
+			"Included config %s can only be partially parsed during dependency graph formation for run --all command as it has a dependency block.",
 			includePath,
 		)
-		return PartialParseConfigFile(includePath, terragruntOptions, includedConfig, decodeList)
+
+		return PartialParseConfigFile(ctx, l, includePath, includedConfig)
 	}
-	return ParseConfigFile(includePath, terragruntOptions, includedConfig, dependencyOutputs)
+
+	config, err := ParseConfigFile(ctx, l, includePath, includedConfig)
+	if err != nil {
+		var configNotFoundError TerragruntConfigNotFoundError
+		if errors.As(err, &configNotFoundError) {
+			return nil, IncludeConfigNotFoundError{
+				IncludePath: includePath,
+				SourcePath:  ctx.TerragruntOptions.TerragruntConfigPath,
+			}
+		}
+
+		return nil, err
+	}
+
+	return config, nil
 }
 
 // handleInclude merges the included config into the current config depending on the merge strategy specified by the
 // user.
-func handleInclude(
-	config *TerragruntConfig,
-	trackInclude *TrackInclude,
-	terragruntOptions *options.TerragruntOptions,
-	dependencyOutputs *cty.Value,
-) (*TerragruntConfig, error) {
-	if trackInclude == nil {
-		return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: HANDLE_INCLUDE_NIL_INCLUDE_CONFIG")
+func handleInclude(ctx *ParsingContext, l log.Logger, config *TerragruntConfig, isPartial bool) (*TerragruntConfig, error) {
+	if ctx.TrackInclude == nil {
+		return nil, errors.New("you reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message.Code: HANDLE_INCLUDE_NIL_INCLUDE_CONFIG")
 	}
 
 	// We merge in the include blocks in reverse order here. The expectation is that the bottom most elements override
 	// those in earlier includes, so we need to merge bottom up instead of top down to ensure this.
-	includeList := trackInclude.CurrentList
+	includeList := ctx.TrackInclude.CurrentList
 	baseConfig := config
+
 	for i := len(includeList) - 1; i >= 0; i-- {
 		includeConfig := includeList[i]
+
 		mergeStrategy, err := includeConfig.GetMergeStrategy()
 		if err != nil {
 			return config, err
 		}
 
-		parsedIncludeConfig, err := parseIncludedConfig(&includeConfig, terragruntOptions, dependencyOutputs, nil)
-		if err != nil {
-			return nil, err
+		var (
+			parsedIncludeConfig *TerragruntConfig
+			logPrefix           string
+		)
+
+		trackFileRead(ctx.FilesRead, includeConfig.Path)
+
+		if isPartial {
+			parsedIncludeConfig, err = partialParseIncludedConfig(ctx, l, &includeConfig)
+			logPrefix = "[Partial] "
+		} else {
+			parsedIncludeConfig, err = parseIncludedConfig(ctx, l, &includeConfig)
 		}
 
-		switch mergeStrategy {
+		if err != nil {
+			return baseConfig, err
+		}
+
+		// TODO: Remove lint suppression
+		switch mergeStrategy { //nolint:exhaustive
 		case NoMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy no merge: not merging config in.", includeConfig.Path)
+			l.Debugf("%sIncluded config %s has strategy no merge: not merging config in.", logPrefix, includeConfig.Path)
 		case ShallowMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy shallow merge: merging config in (shallow).", includeConfig.Path)
-			parsedIncludeConfig.Merge(baseConfig, terragruntOptions)
-			baseConfig = parsedIncludeConfig
-		case DeepMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy deep merge: merging config in (deep).", includeConfig.Path)
-			if err := parsedIncludeConfig.DeepMerge(baseConfig, terragruntOptions); err != nil {
+			l.Debugf("%sIncluded config %s has strategy shallow merge: merging config in (shallow).", logPrefix, includeConfig.Path)
+
+			if err := parsedIncludeConfig.Merge(l, baseConfig, ctx.TerragruntOptions); err != nil {
 				return nil, err
 			}
-			baseConfig = parsedIncludeConfig
-		default:
-			return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: UNKNOWN_MERGE_STRATEGY_%s", mergeStrategy)
-		}
-	}
-	return baseConfig, nil
-}
 
-// handleIncludePartial merges the a partially parsed include config into the child config according to the strategy
-// specified by the user.
-func handleIncludePartial(
-	config *TerragruntConfig,
-	trackInclude *TrackInclude,
-	terragruntOptions *options.TerragruntOptions,
-	decodeList []PartialDecodeSectionType,
-) (*TerragruntConfig, error) {
-	if trackInclude == nil {
-		return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: HANDLE_INCLUDE_PARTIAL_NIL_INCLUDE_CONFIG")
-	}
-
-	// We merge in the include blocks in reverse order here. The expectation is that the bottom most elements override
-	// those in earlier includes, so we need to merge bottom up instead of top down to ensure this.
-	includeList := trackInclude.CurrentList
-	baseConfig := config
-	for i := len(includeList) - 1; i >= 0; i-- {
-		includeConfig := includeList[i]
-		mergeStrategy, err := includeConfig.GetMergeStrategy()
-		if err != nil {
-			return nil, err
-		}
-
-		parsedIncludeConfig, err := partialParseIncludedConfig(&includeConfig, terragruntOptions, decodeList)
-		if err != nil {
-			return nil, err
-		}
-
-		switch mergeStrategy {
-		case NoMerge:
-			terragruntOptions.Logger.Debugf("[Partial] Included config %s has strategy no merge: not merging config in.", includeConfig.Path)
-		case ShallowMerge:
-			terragruntOptions.Logger.Debugf("[Partial] Included config %s has strategy shallow merge: merging config in (shallow).", includeConfig.Path)
-			parsedIncludeConfig.Merge(baseConfig, terragruntOptions)
 			baseConfig = parsedIncludeConfig
 		case DeepMerge:
-			terragruntOptions.Logger.Debugf("[Partial] Included config %s has strategy deep merge: merging config in (deep).", includeConfig.Path)
-			if err := parsedIncludeConfig.DeepMerge(baseConfig, terragruntOptions); err != nil {
+			l.Debugf("%sIncluded config %s has strategy deep merge: merging config in (deep).", logPrefix, includeConfig.Path)
+
+			if err := parsedIncludeConfig.DeepMerge(l, baseConfig, ctx.TerragruntOptions); err != nil {
 				return nil, err
 			}
+
 			baseConfig = parsedIncludeConfig
 		default:
-			return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: UNKNOWN_MERGE_STRATEGY_%s_PARTIAL", mergeStrategy)
+			return nil, fmt.Errorf("you reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: UNKNOWN_MERGE_STRATEGY_%s", mergeStrategy)
 		}
 	}
+
 	return baseConfig, nil
 }
 
@@ -193,49 +181,53 @@ func handleIncludePartial(
 // dependency block configurations between the included config and the child config. This allows us to merge the two
 // dependencies prior to retrieving the outputs, allowing you to have partial configuration that is overridden by a
 // child.
-func handleIncludeForDependency(
-	childDecodedDependency terragruntDependency,
-	trackInclude *TrackInclude,
-	terragruntOptions *options.TerragruntOptions,
-) (*terragruntDependency, error) {
-	if trackInclude == nil {
-		return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: HANDLE_INCLUDE_DEPENDENCY_NIL_INCLUDE_CONFIG")
+func handleIncludeForDependency(ctx *ParsingContext, l log.Logger, childDecodedDependency TerragruntDependency) (*TerragruntDependency, error) {
+	if ctx.TrackInclude == nil {
+		return nil, errors.New("you reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: HANDLE_INCLUDE_DEPENDENCY_NIL_INCLUDE_CONFIG")
 	}
 	// We merge in the include blocks in reverse order here. The expectation is that the bottom most elements override
 	// those in earlier includes, so we need to merge bottom up instead of top down to ensure this.
-	includeList := trackInclude.CurrentList
+	includeList := ctx.TrackInclude.CurrentList
 	baseDependencyBlock := childDecodedDependency.Dependencies
+
 	for i := len(includeList) - 1; i >= 0; i-- {
 		includeConfig := includeList[i]
+
 		mergeStrategy, err := includeConfig.GetMergeStrategy()
 		if err != nil {
 			return nil, err
 		}
 
-		includedPartialParse, err := partialParseIncludedConfig(&includeConfig, terragruntOptions, []PartialDecodeSectionType{DependencyBlock})
+		includedPartialParse, err := partialParseIncludedConfig(
+			ctx.WithDecodeList(DependencyBlock, FeatureFlagsBlock, ExcludeBlock, ErrorsBlock), l, &includeConfig)
 		if err != nil {
 			return nil, err
 		}
 
-		switch mergeStrategy {
+		// TODO: Remove lint suppression
+		switch mergeStrategy { //nolint:exhaustive
 		case NoMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy no merge: not merging config in for dependency.", includeConfig.Path)
+			l.Debugf("Included config %s has strategy no merge: not merging config in for dependency.", includeConfig.Path)
 		case ShallowMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy shallow merge: merging config in (shallow) for dependency.", includeConfig.Path)
+			l.Debugf("Included config %s has strategy shallow merge: merging config in (shallow) for dependency.", includeConfig.Path)
+
 			mergedDependencyBlock := mergeDependencyBlocks(includedPartialParse.TerragruntDependencies, baseDependencyBlock)
 			baseDependencyBlock = mergedDependencyBlock
 		case DeepMerge:
-			terragruntOptions.Logger.Debugf("Included config %s has strategy deep merge: merging config in (deep) for dependency.", includeConfig.Path)
+			l.Debugf("Included config %s has strategy deep merge: merging config in (deep) for dependency.", includeConfig.Path)
+
 			mergedDependencyBlock, err := deepMergeDependencyBlocks(includedPartialParse.TerragruntDependencies, baseDependencyBlock)
 			if err != nil {
 				return nil, err
 			}
+
 			baseDependencyBlock = mergedDependencyBlock
 		default:
-			return nil, fmt.Errorf("You reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: UNKNOWN_MERGE_STRATEGY_%s_DEPENDENCY", mergeStrategy)
+			return nil, fmt.Errorf("you reached an impossible condition. This is most likely a bug in terragrunt. Please open an issue at github.com/gruntwork-io/terragrunt with this error message. Code: UNKNOWN_MERGE_STRATEGY_%s_DEPENDENCY", mergeStrategy)
 		}
 	}
-	return &terragruntDependency{Dependencies: baseDependencyBlock}, nil
+
+	return &TerragruntDependency{Dependencies: baseDependencyBlock}, nil
 }
 
 // Merge performs a shallow merge of the given sourceConfig into the targetConfig. sourceConfig will override common
@@ -247,94 +239,112 @@ func handleIncludeForDependency(
 // NOTE: dependencies block is a special case and is merged deeply. This is necessary to ensure the configstack system
 // works correctly, as it uses the `Dependencies` list to track the dependencies of modules for graph building purposes.
 // This list includes the dependencies added from dependency blocks, which is handled in a different stage.
-func (targetConfig *TerragruntConfig) Merge(sourceConfig *TerragruntConfig, terragruntOptions *options.TerragruntOptions) {
+func (cfg *TerragruntConfig) Merge(l log.Logger, sourceConfig *TerragruntConfig, terragruntOptions *options.TerragruntOptions) error {
 	// Merge simple attributes first
 	if sourceConfig.DownloadDir != "" {
-		targetConfig.DownloadDir = sourceConfig.DownloadDir
+		cfg.DownloadDir = sourceConfig.DownloadDir
 	}
 
 	if sourceConfig.IamRole != "" {
-		targetConfig.IamRole = sourceConfig.IamRole
+		cfg.IamRole = sourceConfig.IamRole
 	}
 
 	if sourceConfig.IamAssumeRoleDuration != nil {
-		targetConfig.IamAssumeRoleDuration = sourceConfig.IamAssumeRoleDuration
+		cfg.IamAssumeRoleDuration = sourceConfig.IamAssumeRoleDuration
+	}
+
+	if sourceConfig.IamAssumeRoleSessionName != "" {
+		cfg.IamAssumeRoleSessionName = sourceConfig.IamAssumeRoleSessionName
+	}
+
+	if sourceConfig.IamWebIdentityToken != "" {
+		cfg.IamWebIdentityToken = sourceConfig.IamWebIdentityToken
 	}
 
 	if sourceConfig.TerraformVersionConstraint != "" {
-		targetConfig.TerraformVersionConstraint = sourceConfig.TerraformVersionConstraint
+		cfg.TerraformVersionConstraint = sourceConfig.TerraformVersionConstraint
 	}
 
 	if sourceConfig.TerraformBinary != "" {
-		targetConfig.TerraformBinary = sourceConfig.TerraformBinary
+		cfg.TerraformBinary = sourceConfig.TerraformBinary
 	}
 
 	if sourceConfig.PreventDestroy != nil {
-		targetConfig.PreventDestroy = sourceConfig.PreventDestroy
-	}
-
-	if sourceConfig.RetryMaxAttempts != nil {
-		targetConfig.RetryMaxAttempts = sourceConfig.RetryMaxAttempts
-	}
-
-	if sourceConfig.RetrySleepIntervalSec != nil {
-		targetConfig.RetrySleepIntervalSec = sourceConfig.RetrySleepIntervalSec
+		cfg.PreventDestroy = sourceConfig.PreventDestroy
 	}
 
 	if sourceConfig.TerragruntVersionConstraint != "" {
-		targetConfig.TerragruntVersionConstraint = sourceConfig.TerragruntVersionConstraint
+		cfg.TerragruntVersionConstraint = sourceConfig.TerragruntVersionConstraint
 	}
 
-	// Skip has to be set specifically in each file that should be skipped
-	targetConfig.Skip = sourceConfig.Skip
+	if sourceConfig.Engine != nil {
+		cfg.Engine = sourceConfig.Engine.Clone()
+	}
+
+	if sourceConfig.Exclude != nil {
+		cfg.Exclude = sourceConfig.Exclude.Clone()
+	}
+
+	if sourceConfig.Errors != nil {
+		cfg.Errors = sourceConfig.Errors.Clone()
+	}
 
 	if sourceConfig.RemoteState != nil {
-		targetConfig.RemoteState = sourceConfig.RemoteState
+		cfg.RemoteState = sourceConfig.RemoteState
 	}
 
 	if sourceConfig.Terraform != nil {
-		if targetConfig.Terraform == nil {
-			targetConfig.Terraform = sourceConfig.Terraform
+		if cfg.Terraform == nil {
+			cfg.Terraform = sourceConfig.Terraform
 		} else {
 			if sourceConfig.Terraform.Source != nil {
-				targetConfig.Terraform.Source = sourceConfig.Terraform.Source
+				cfg.Terraform.Source = sourceConfig.Terraform.Source
 			}
-			mergeExtraArgs(terragruntOptions, sourceConfig.Terraform.ExtraArgs, &targetConfig.Terraform.ExtraArgs)
 
-			mergeHooks(terragruntOptions, sourceConfig.Terraform.BeforeHooks, &targetConfig.Terraform.BeforeHooks)
-			mergeHooks(terragruntOptions, sourceConfig.Terraform.AfterHooks, &targetConfig.Terraform.AfterHooks)
-			mergeErrorHooks(terragruntOptions, sourceConfig.Terraform.ErrorHooks, &targetConfig.Terraform.ErrorHooks)
+			if sourceConfig.Terraform.CopyTerraformLockFile != nil {
+				cfg.Terraform.CopyTerraformLockFile = sourceConfig.Terraform.CopyTerraformLockFile
+			}
+
+			mergeExtraArgs(l, sourceConfig.Terraform.ExtraArgs, &cfg.Terraform.ExtraArgs)
+
+			mergeHooks(l, sourceConfig.Terraform.BeforeHooks, &cfg.Terraform.BeforeHooks)
+			mergeHooks(l, sourceConfig.Terraform.AfterHooks, &cfg.Terraform.AfterHooks)
+			mergeErrorHooks(l, sourceConfig.Terraform.ErrorHooks, &cfg.Terraform.ErrorHooks)
 		}
 	}
 
 	// Dependency blocks are shallow merged by name
-	targetConfig.TerragruntDependencies = mergeDependencyBlocks(targetConfig.TerragruntDependencies, sourceConfig.TerragruntDependencies)
+	cfg.TerragruntDependencies = mergeDependencyBlocks(cfg.TerragruntDependencies, sourceConfig.TerragruntDependencies)
+
+	cfg.FeatureFlags = mergeFeatureFlags(cfg.FeatureFlags, sourceConfig.FeatureFlags)
 
 	// Deep merge the dependencies list. This is different from dependency blocks, and refers to the deprecated
 	// dependencies block!
 	if sourceConfig.Dependencies != nil {
-		if targetConfig.Dependencies == nil {
-			targetConfig.Dependencies = sourceConfig.Dependencies
+		if cfg.Dependencies == nil {
+			cfg.Dependencies = sourceConfig.Dependencies
 		} else {
-			targetConfig.Dependencies.Merge(sourceConfig.Dependencies)
+			cfg.Dependencies.Merge(sourceConfig.Dependencies)
 		}
-	}
-
-	if sourceConfig.RetryableErrors != nil {
-		targetConfig.RetryableErrors = sourceConfig.RetryableErrors
 	}
 
 	// Merge the generate configs. This is a shallow merge. Meaning, if the child has the same name generate block, then the
 	// child's generate block will override the parent's block.
-	for key, val := range sourceConfig.GenerateConfigs {
-		targetConfig.GenerateConfigs[key] = val
+
+	err := validateGenerateConfigs(&sourceConfig.GenerateConfigs, &cfg.GenerateConfigs)
+	if err != nil {
+		return err
 	}
+
+	maps.Copy(cfg.GenerateConfigs, sourceConfig.GenerateConfigs)
 
 	if sourceConfig.Inputs != nil {
-		targetConfig.Inputs = mergeInputs(sourceConfig.Inputs, targetConfig.Inputs)
+		cfg.Inputs = mergeInputs(sourceConfig.Inputs, cfg.Inputs)
 	}
 
-	copyFieldsMetadata(sourceConfig, targetConfig)
+	CopyFieldsMetadata(sourceConfig, cfg)
+
+	return nil
 }
 
 // DeepMerge performs a deep merge of the given sourceConfig into the targetConfig. Deep merge is defined as follows:
@@ -351,55 +361,78 @@ func (targetConfig *TerragruntConfig) Merge(sourceConfig *TerragruntConfig, terr
 //   - dependency blocks (TerragruntDependencies) [These blocks need to retrieve outputs, so we need to merge during
 //     the parsing step, not after the full config is decoded]
 //   - locals [These blocks are not merged by design]
-func (targetConfig *TerragruntConfig) DeepMerge(sourceConfig *TerragruntConfig, terragruntOptions *options.TerragruntOptions) error {
+func (cfg *TerragruntConfig) DeepMerge(l log.Logger, sourceConfig *TerragruntConfig, terragruntOptions *options.TerragruntOptions) error {
 	// Merge simple attributes first
 	if sourceConfig.DownloadDir != "" {
-		targetConfig.DownloadDir = sourceConfig.DownloadDir
+		cfg.DownloadDir = sourceConfig.DownloadDir
 	}
 
 	if sourceConfig.IamRole != "" {
-		targetConfig.IamRole = sourceConfig.IamRole
+		cfg.IamRole = sourceConfig.IamRole
 	}
 
 	if sourceConfig.IamAssumeRoleDuration != nil {
-		targetConfig.IamAssumeRoleDuration = sourceConfig.IamAssumeRoleDuration
+		cfg.IamAssumeRoleDuration = sourceConfig.IamAssumeRoleDuration
+	}
+
+	if sourceConfig.IamAssumeRoleSessionName != "" {
+		cfg.IamAssumeRoleSessionName = sourceConfig.IamAssumeRoleSessionName
+	}
+
+	if sourceConfig.IamWebIdentityToken != "" {
+		cfg.IamWebIdentityToken = sourceConfig.IamWebIdentityToken
 	}
 
 	if sourceConfig.TerraformVersionConstraint != "" {
-		targetConfig.TerraformVersionConstraint = sourceConfig.TerraformVersionConstraint
+		cfg.TerraformVersionConstraint = sourceConfig.TerraformVersionConstraint
 	}
 
 	if sourceConfig.TerraformBinary != "" {
-		targetConfig.TerraformBinary = sourceConfig.TerraformBinary
+		cfg.TerraformBinary = sourceConfig.TerraformBinary
 	}
 
 	if sourceConfig.PreventDestroy != nil {
-		targetConfig.PreventDestroy = sourceConfig.PreventDestroy
-	}
-
-	if sourceConfig.RetryMaxAttempts != nil {
-		targetConfig.RetryMaxAttempts = sourceConfig.RetryMaxAttempts
-	}
-
-	if sourceConfig.RetrySleepIntervalSec != nil {
-		targetConfig.RetrySleepIntervalSec = sourceConfig.RetrySleepIntervalSec
+		cfg.PreventDestroy = sourceConfig.PreventDestroy
 	}
 
 	if sourceConfig.TerragruntVersionConstraint != "" {
-		targetConfig.TerragruntVersionConstraint = sourceConfig.TerragruntVersionConstraint
+		cfg.TerragruntVersionConstraint = sourceConfig.TerragruntVersionConstraint
 	}
 
-	// Skip has to be set specifically in each file that should be skipped
-	targetConfig.Skip = sourceConfig.Skip
+	if sourceConfig.Engine != nil {
+		if cfg.Engine == nil {
+			cfg.Engine = &EngineConfig{}
+		}
+
+		cfg.Engine.Merge(sourceConfig.Engine)
+	}
+
+	if sourceConfig.Exclude != nil {
+		if cfg.Exclude == nil {
+			cfg.Exclude = &ExcludeConfig{}
+		}
+
+		cfg.Exclude.Merge(sourceConfig.Exclude)
+	}
+
+	if sourceConfig.Errors != nil {
+		if cfg.Errors == nil {
+			cfg.Errors = &ErrorsConfig{}
+		}
+
+		cfg.Errors.Merge(sourceConfig.Errors)
+	}
 
 	// Copy only dependencies which doesn't exist in source
 	if sourceConfig.Dependencies != nil {
 		resultModuleDependencies := &ModuleDependencies{}
-		if targetConfig.Dependencies != nil {
+
+		if cfg.Dependencies != nil {
 			// take in result dependencies only paths which aren't defined in source
 			// Fix for issue: https://github.com/gruntwork-io/terragrunt/issues/1900
-			targetPathMap := fetchDependencyPaths(targetConfig)
+			targetPathMap := fetchDependencyPaths(cfg)
 			sourcePathMap := fetchDependencyPaths(sourceConfig)
+
 			for key, value := range targetPathMap {
 				_, found := sourcePathMap[key]
 				if !found {
@@ -408,94 +441,153 @@ func (targetConfig *TerragruntConfig) DeepMerge(sourceConfig *TerragruntConfig, 
 			}
 			// copy target paths which are defined only in Dependencies and not in TerragruntDependencies
 			// if TerragruntDependencies will be empty, all targetConfig.Dependencies.Paths will be copied to resultModuleDependencies.Paths
-			for _, dependencyPath := range targetConfig.Dependencies.Paths {
+			for _, dependencyPath := range cfg.Dependencies.Paths {
 				var addPath = true
+
 				for _, targetPath := range targetPathMap {
 					if dependencyPath == targetPath { // path already defined in TerragruntDependencies, skip adding
 						addPath = false
 						break
 					}
 				}
+
 				if addPath {
 					resultModuleDependencies.Paths = append(resultModuleDependencies.Paths, dependencyPath)
 				}
 			}
 		}
+
 		resultModuleDependencies.Paths = append(resultModuleDependencies.Paths, sourceConfig.Dependencies.Paths...)
-		targetConfig.Dependencies = resultModuleDependencies
+		cfg.Dependencies = resultModuleDependencies
 	}
 
 	// Dependency blocks are deep merged by name
-	mergedDeps, err := deepMergeDependencyBlocks(targetConfig.TerragruntDependencies, sourceConfig.TerragruntDependencies)
+	mergedDeps, err := deepMergeDependencyBlocks(cfg.TerragruntDependencies, sourceConfig.TerragruntDependencies)
 	if err != nil {
 		return err
 	}
-	targetConfig.TerragruntDependencies = mergedDeps
 
-	if sourceConfig.RetryableErrors != nil {
-		targetConfig.RetryableErrors = append(targetConfig.RetryableErrors, sourceConfig.RetryableErrors...)
+	cfg.TerragruntDependencies = mergedDeps
+
+	mergedFlags, err := deepMergeFeatureBlocks(cfg.FeatureFlags, sourceConfig.FeatureFlags)
+	if err != nil {
+		return err
 	}
+
+	cfg.FeatureFlags = mergedFlags
 
 	// Handle complex structs by recursively merging the structs together
 	if sourceConfig.Terraform != nil {
-		if targetConfig.Terraform == nil {
-			targetConfig.Terraform = sourceConfig.Terraform
+		if cfg.Terraform == nil {
+			cfg.Terraform = sourceConfig.Terraform
 		} else {
 			if sourceConfig.Terraform.Source != nil {
-				targetConfig.Terraform.Source = sourceConfig.Terraform.Source
+				cfg.Terraform.Source = sourceConfig.Terraform.Source
+			}
+
+			if sourceConfig.Terraform.CopyTerraformLockFile != nil {
+				cfg.Terraform.CopyTerraformLockFile = sourceConfig.Terraform.CopyTerraformLockFile
 			}
 
 			if sourceConfig.Terraform.IncludeInCopy != nil {
 				srcList := *sourceConfig.Terraform.IncludeInCopy
-				if targetConfig.Terraform.IncludeInCopy != nil {
-					targetList := *targetConfig.Terraform.IncludeInCopy
+
+				if cfg.Terraform.IncludeInCopy != nil {
+					targetList := *cfg.Terraform.IncludeInCopy
 					combinedList := append(srcList, targetList...)
-					targetConfig.Terraform.IncludeInCopy = &combinedList
+					cfg.Terraform.IncludeInCopy = &combinedList
 				} else {
-					targetConfig.Terraform.IncludeInCopy = &srcList
+					cfg.Terraform.IncludeInCopy = &srcList
 				}
 			}
 
-			mergeExtraArgs(terragruntOptions, sourceConfig.Terraform.ExtraArgs, &targetConfig.Terraform.ExtraArgs)
+			if sourceConfig.Terraform.ExcludeFromCopy != nil {
+				srcList := *sourceConfig.Terraform.ExcludeFromCopy
 
-			mergeHooks(terragruntOptions, sourceConfig.Terraform.BeforeHooks, &targetConfig.Terraform.BeforeHooks)
-			mergeHooks(terragruntOptions, sourceConfig.Terraform.AfterHooks, &targetConfig.Terraform.AfterHooks)
-			mergeErrorHooks(terragruntOptions, sourceConfig.Terraform.ErrorHooks, &targetConfig.Terraform.ErrorHooks)
+				if cfg.Terraform.ExcludeFromCopy != nil {
+					targetList := *cfg.Terraform.ExcludeFromCopy
+					combinedList := append(srcList, targetList...)
+					cfg.Terraform.ExcludeFromCopy = &combinedList
+				} else {
+					cfg.Terraform.ExcludeFromCopy = &srcList
+				}
+			}
+
+			mergeExtraArgs(l, sourceConfig.Terraform.ExtraArgs, &cfg.Terraform.ExtraArgs)
+
+			mergeHooks(l, sourceConfig.Terraform.BeforeHooks, &cfg.Terraform.BeforeHooks)
+			mergeHooks(l, sourceConfig.Terraform.AfterHooks, &cfg.Terraform.AfterHooks)
+			mergeErrorHooks(l, sourceConfig.Terraform.ErrorHooks, &cfg.Terraform.ErrorHooks)
 		}
 	}
 
 	if sourceConfig.Inputs != nil {
-		mergedInputs, err := deepMergeInputs(sourceConfig.Inputs, targetConfig.Inputs)
+		mergedInputs, err := deepMergeInputs(sourceConfig.Inputs, cfg.Inputs)
 		if err != nil {
 			return err
 		}
-		targetConfig.Inputs = mergedInputs
+
+		cfg.Inputs = mergedInputs
 	}
 
 	// MAINTAINER'S NOTE: The following structs cannot be deep merged due to an implementation detail (they do not
 	// support nil attributes, so we can't determine if an attribute was intentionally set, or was defaulted from
 	// unspecified - this is especially problematic for bool attributes).
 	if sourceConfig.RemoteState != nil {
-		targetConfig.RemoteState = sourceConfig.RemoteState
-	}
-	for key, val := range sourceConfig.GenerateConfigs {
-		targetConfig.GenerateConfigs[key] = val
+		cfg.RemoteState = sourceConfig.RemoteState
 	}
 
-	copyFieldsMetadata(sourceConfig, targetConfig)
+	maps.Copy(cfg.GenerateConfigs, sourceConfig.GenerateConfigs)
+
+	CopyFieldsMetadata(sourceConfig, cfg)
+
 	return nil
 }
 
-// fetchDependencyMap - return from configuration map with dependency_name: path
+// fetchDependencyPaths - return from configuration map with dependency_name: path
 func fetchDependencyPaths(config *TerragruntConfig) map[string]string {
 	var m = make(map[string]string)
 	if config == nil {
 		return m
 	}
+
 	for _, dependency := range config.TerragruntDependencies {
-		m[dependency.Name] = dependency.ConfigPath
+		m[dependency.Name] = dependency.ConfigPath.AsString()
 	}
+
 	return m
+}
+
+// merge feature flags by name.
+func mergeFeatureFlags(targetFlags []*FeatureFlag, sourceFlags []*FeatureFlag) []*FeatureFlag {
+	if sourceFlags == nil && targetFlags == nil {
+		return nil
+	}
+
+	keys := make([]string, 0, len(targetFlags))
+
+	flagBlocks := make(map[string]*FeatureFlag)
+	for _, flags := range targetFlags {
+		flagBlocks[flags.Name] = flags
+		keys = append(keys, flags.Name)
+	}
+
+	for _, dep := range sourceFlags {
+		_, hasSameKey := flagBlocks[dep.Name]
+		if !hasSameKey {
+			keys = append(keys, dep.Name)
+		}
+
+		flagBlocks[dep.Name] = dep
+	}
+
+	combinedFlags := make([]*FeatureFlag, 0, len(keys))
+
+	for _, key := range keys {
+		combinedFlags = append(combinedFlags, flagBlocks[key])
+	}
+
+	return combinedFlags
 }
 
 // Merge dependency blocks shallowly. If the source list has the same name as the target, it will override the
@@ -511,6 +603,7 @@ func mergeDependencyBlocks(targetDependencies []Dependency, sourceDependencies [
 		dependencyBlocks[dep.Name] = dep
 		keys = append(keys, dep.Name)
 	}
+
 	for _, dep := range sourceDependencies {
 		_, hasSameKey := dependencyBlocks[dep.Name]
 		if !hasSameKey {
@@ -524,6 +617,7 @@ func mergeDependencyBlocks(targetDependencies []Dependency, sourceDependencies [
 	for _, key := range keys {
 		combinedDeps = append(combinedDeps, dependencyBlocks[key])
 	}
+
 	return combinedDeps
 }
 
@@ -540,6 +634,7 @@ func deepMergeDependencyBlocks(targetDependencies []Dependency, sourceDependenci
 		dependencyBlocks[dep.Name] = dep
 		keys = append(keys, dep.Name)
 	}
+
 	for _, dep := range sourceDependencies {
 		sameKeyDep, hasSameKey := dependencyBlocks[dep.Name]
 		if hasSameKey {
@@ -547,18 +642,20 @@ func deepMergeDependencyBlocks(targetDependencies []Dependency, sourceDependenci
 			if err := sameKeyDepPtr.DeepMerge(dep); err != nil {
 				return nil, err
 			}
+
 			dependencyBlocks[dep.Name] = *sameKeyDepPtr
 		} else {
 			dependencyBlocks[dep.Name] = dep
 			keys = append(keys, dep.Name)
 		}
-
 	}
+
 	// Now convert the map to list and set target
 	combinedDeps := []Dependency{}
 	for _, key := range keys {
 		combinedDeps = append(combinedDeps, dependencyBlocks[key])
 	}
+
 	return combinedDeps, nil
 }
 
@@ -572,14 +669,14 @@ func deepMergeDependencyBlocks(targetDependencies []Dependency, sourceDependenci
 // extra_arguments on the terraform cli.
 // Therefore, if .tfvar files from both the parent and child contain a variable
 // with the same name, the value from the child will win.
-func mergeExtraArgs(terragruntOptions *options.TerragruntOptions, childExtraArgs []TerraformExtraArguments, parentExtraArgs *[]TerraformExtraArguments) {
+func mergeExtraArgs(l log.Logger, childExtraArgs []TerraformExtraArguments, parentExtraArgs *[]TerraformExtraArguments) {
 	result := *parentExtraArgs
 	for _, child := range childExtraArgs {
 		parentExtraArgsWithSameName := getIndexOfExtraArgsWithName(result, child.Name)
 		if parentExtraArgsWithSameName != -1 {
 			// If the parent contains an extra_arguments with the same name as the child,
 			// then override the parent's extra_arguments with the child's.
-			terragruntOptions.Logger.Debugf("extra_arguments '%v' from child overriding parent", child.Name)
+			l.Debugf("extra_arguments '%v' from child overriding parent", child.Name)
 			result[parentExtraArgsWithSameName] = child
 		} else {
 			// If the parent does not contain an extra_arguments with the same name as the child
@@ -588,31 +685,27 @@ func mergeExtraArgs(terragruntOptions *options.TerragruntOptions, childExtraArgs
 			result = append(result, child)
 		}
 	}
+
 	*parentExtraArgs = result
 }
 
-func mergeInputs(childInputs map[string]interface{}, parentInputs map[string]interface{}) map[string]interface{} {
-	out := map[string]interface{}{}
+func mergeInputs(childInputs map[string]any, parentInputs map[string]any) map[string]any {
+	out := map[string]any{}
 
-	for key, value := range parentInputs {
-		out[key] = value
-	}
+	maps.Copy(out, parentInputs)
 
-	for key, value := range childInputs {
-		out[key] = value
-	}
+	maps.Copy(out, childInputs)
 
 	return out
 }
 
-func deepMergeInputs(childInputs map[string]interface{}, parentInputs map[string]interface{}) (map[string]interface{}, error) {
-	out := map[string]interface{}{}
-	for key, value := range parentInputs {
-		out[key] = value
-	}
+func deepMergeInputs(childInputs map[string]any, parentInputs map[string]any) (map[string]any, error) {
+	out := map[string]any{}
+	maps.Copy(out, parentInputs)
 
 	err := mergo.Merge(&out, childInputs, mergo.WithAppendSlice, mergo.WithOverride)
-	return out, errors.WithStackTrace(err)
+
+	return out, errors.New(err)
 }
 
 // Merge the hooks (before_hook and after_hook).
@@ -622,14 +715,14 @@ func deepMergeInputs(childInputs map[string]interface{}, parentInputs map[string
 // If a child's hook has a different name from all of the parent's hooks,
 // then the child's hook will be added to the end of the parent's.
 // Therefore, the child with the same name overrides the parent
-func mergeHooks(terragruntOptions *options.TerragruntOptions, childHooks []Hook, parentHooks *[]Hook) {
+func mergeHooks(l log.Logger, childHooks []Hook, parentHooks *[]Hook) {
 	result := *parentHooks
 	for _, child := range childHooks {
 		parentHookWithSameName := getIndexOfHookWithName(result, child.Name)
 		if parentHookWithSameName != -1 {
 			// If the parent contains a hook with the same name as the child,
 			// then override the parent's hook with the child's.
-			terragruntOptions.Logger.Debugf("hook '%v' from child overriding parent", child.Name)
+			l.Debugf("hook '%v' from child overriding parent", child.Name)
 			result[parentHookWithSameName] = child
 		} else {
 			// If the parent does not contain a hook with the same name as the child
@@ -637,20 +730,21 @@ func mergeHooks(terragruntOptions *options.TerragruntOptions, childHooks []Hook,
 			result = append(result, child)
 		}
 	}
+
 	*parentHooks = result
 }
 
 // Merge the error hooks (error_hook).
 // Does the same thing as mergeHooks but for error hooks
 // TODO: Figure out more DRY way to do this
-func mergeErrorHooks(terragruntOptions *options.TerragruntOptions, childHooks []ErrorHook, parentHooks *[]ErrorHook) {
+func mergeErrorHooks(l log.Logger, childHooks []ErrorHook, parentHooks *[]ErrorHook) {
 	result := *parentHooks
 	for _, child := range childHooks {
 		parentHookWithSameName := getIndexOfErrorHookWithName(result, child.Name)
 		if parentHookWithSameName != -1 {
 			// If the parent contains a hook with the same name as the child,
 			// then override the parent's hook with the child's.
-			terragruntOptions.Logger.Debugf("hook '%v' from child overriding parent", child.Name)
+			l.Debugf("hook '%v' from child overriding parent", child.Name)
 			result[parentHookWithSameName] = child
 		} else {
 			// If the parent does not contain a hook with the same name as the child
@@ -658,51 +752,48 @@ func mergeErrorHooks(terragruntOptions *options.TerragruntOptions, childHooks []
 			result = append(result, child)
 		}
 	}
+
 	*parentHooks = result
 }
 
 // getTrackInclude converts the terragrunt include blocks into TrackInclude structs that differentiate between an
-// included config in the current parsing context, and an included config that was passed through from a previous
-// parsing context.
-func getTrackInclude(
-	terragruntIncludeList []IncludeConfig,
-	includeFromChild *IncludeConfig,
-	terragruntOptions *options.TerragruntOptions,
-) (*TrackInclude, error) {
+// included config in the current parsing ctx, and an included config that was passed through from a previous
+// parsing ctx.
+func getTrackInclude(ctx *ParsingContext, terragruntIncludeList IncludeConfigs, includeFromChild *IncludeConfig) (*TrackInclude, error) {
 	includedPaths := []string{}
 	terragruntIncludeMap := make(map[string]IncludeConfig, len(terragruntIncludeList))
+
 	for _, tgInc := range terragruntIncludeList {
 		includedPaths = append(includedPaths, tgInc.Path)
 		terragruntIncludeMap[tgInc.Name] = tgInc
 	}
 
 	hasInclude := len(terragruntIncludeList) > 0
-	if hasInclude && includeFromChild != nil {
+
+	trackInc := TrackInclude{
+		CurrentList: terragruntIncludeList,
+		CurrentMap:  terragruntIncludeMap,
+	}
+
+	switch {
+	case hasInclude && includeFromChild != nil:
 		// tgInc appears in a parent that is already included, which means a nested include block. This is not
 		// something we currently support.
-		err := errors.WithStackTrace(TooManyLevelsOfInheritance{
-			ConfigPath:             terragruntOptions.TerragruntConfigPath,
+		err := errors.New(TooManyLevelsOfInheritanceError{
+			ConfigPath:             ctx.TerragruntOptions.TerragruntConfigPath,
 			FirstLevelIncludePath:  includeFromChild.Path,
 			SecondLevelIncludePath: strings.Join(includedPaths, ","),
 		})
-		return nil, err
-	} else if hasInclude && includeFromChild == nil {
-		// Current parsing context where there is no included config already loaded.
-		trackInc := TrackInclude{
-			CurrentList: terragruntIncludeList,
-			CurrentMap:  terragruntIncludeMap,
-			Original:    nil,
-		}
-		return &trackInc, nil
-	} else {
-		// Parsing context where there is an included config already loaded.
-		trackInc := TrackInclude{
-			CurrentList: terragruntIncludeList,
-			CurrentMap:  terragruntIncludeMap,
-			Original:    includeFromChild,
-		}
-		return &trackInc, nil
+
+		return &TrackInclude{}, err
+	case hasInclude && includeFromChild == nil:
+		// Current parsing ctx where there is no included config already loaded.
+	case !hasInclude:
+		// Parsing ctx where there is an included config already loaded.
+		trackInc.Original = includeFromChild
 	}
+
+	return &trackInc, nil
 }
 
 // updateBareIncludeBlock searches the parsed terragrunt contents for a bare include block (include without a label),
@@ -710,27 +801,51 @@ func getTrackInclude(
 // label counts when parsing out labels with a go struct.
 //
 // Returns the updated contents, a boolean indicated whether anything changed, and an error (if any).
-func updateBareIncludeBlock(file *hcl.File, filename string) ([]byte, bool, error) {
-	if filepath.Ext(filename) == ".json" {
-		return updateBareIncludeBlockJSON(file.Bytes)
+func updateBareIncludeBlock(file *hclparse.File) error {
+	// To save us from doing a lot of extra work, first going to check to see if the file has a naked include, first.
+	// If it doesn't, we aren't going to bother fully parsing the file.
+	if !detectBareIncludeUsage(file) {
+		return nil
 	}
 
-	hclFile, diags := hclwrite.ParseConfig(file.Bytes, filename, hcl.InitialPos)
-	if diags.HasErrors() {
-		return nil, false, errors.WithStackTrace(diags)
-	}
+	var (
+		codeWasUpdated bool
+		content        []byte
+		err            error
+	)
 
-	codeWasUpdated := false
-	for _, block := range hclFile.Body().Blocks() {
-		if block.Type() == "include" && len(block.Labels()) == 0 {
-			if codeWasUpdated {
-				return nil, false, errors.WithStackTrace(MultipleBareIncludeBlocksErr{})
-			}
-			block.SetLabels([]string{bareIncludeKey})
-			codeWasUpdated = true
+	switch filepath.Ext(file.ConfigPath) {
+	case ".json":
+		content, codeWasUpdated, err = updateBareIncludeBlockJSON(file.Bytes)
+		if err != nil {
+			return err
 		}
+	default:
+		hclFile, diags := hclwrite.ParseConfig(file.Bytes, file.ConfigPath, hcl.InitialPos)
+		if diags.HasErrors() {
+			return errors.New(diags)
+		}
+
+		for _, block := range hclFile.Body().Blocks() {
+			if block.Type() == MetadataInclude && len(block.Labels()) == 0 {
+				if codeWasUpdated {
+					return errors.New(MultipleBareIncludeBlocksErr{})
+				}
+
+				block.SetLabels([]string{bareIncludeKey})
+
+				codeWasUpdated = true
+			}
+		}
+
+		content = hclFile.Bytes()
 	}
-	return hclFile.Bytes(), codeWasUpdated, nil
+
+	if !codeWasUpdated {
+		return nil
+	}
+
+	return file.Update(content)
 }
 
 // updateBareIncludeBlockJSON implements the logic for updateBareIncludeBlock when the terragrunt.hcl configuration is
@@ -778,17 +893,19 @@ func updateBareIncludeBlock(file *hcl.File, filename string) ([]byte, bool, erro
 // If the multiple include blocks are encoded in this way in the json configuration, nothing needs to be done by this
 // function.
 func updateBareIncludeBlockJSON(fileBytes []byte) ([]byte, bool, error) {
-	var parsed map[string]interface{}
+	var parsed map[string]any
 	if err := json.Unmarshal(fileBytes, &parsed); err != nil {
-		return nil, false, errors.WithStackTrace(err)
+		return nil, false, errors.New(err)
 	}
-	includeBlock, hasKey := parsed["include"]
+
+	includeBlock, hasKey := parsed[MetadataInclude]
 	if !hasKey {
 		// No include block, so don't do anything
 		return fileBytes, false, nil
 	}
+
 	switch typed := includeBlock.(type) {
-	case []interface{}:
+	case []any:
 		if len(typed) == 0 {
 			// No include block, so don't do anything
 			return nil, false, nil
@@ -796,7 +913,7 @@ func updateBareIncludeBlockJSON(fileBytes []byte) ([]byte, bool, error) {
 			// Could be multiple bare includes, or Case 3. We simplify the handling of this case by erroring out,
 			// ignoring the possibility of Case 3, which, while valid HCL encoding, is too complex to detect and handle
 			// here. Instead we will recommend users use the object encoding.
-			return nil, false, errors.WithStackTrace(MultipleBareIncludeBlocksErr{})
+			return nil, false, errors.New(MultipleBareIncludeBlocksErr{})
 		}
 
 		// Make sure this is Case 2, and not Case 3 with a single labeled block. If Case 2, update to inject the labeled
@@ -805,8 +922,9 @@ func updateBareIncludeBlockJSON(fileBytes []byte) ([]byte, bool, error) {
 		if jsonIsIncludeBlock(singleBlock) {
 			return updateSingleBareIncludeInParsedJSON(parsed, singleBlock)
 		}
+
 		return nil, false, nil
-	case map[string]interface{}:
+	case map[string]any:
 		if len(typed) == 0 {
 			// No include block, so don't do anything
 			return nil, false, nil
@@ -817,19 +935,21 @@ func updateBareIncludeBlockJSON(fileBytes []byte) ([]byte, bool, error) {
 		if jsonIsIncludeBlock(typed) {
 			return updateSingleBareIncludeInParsedJSON(parsed, typed)
 		}
+
 		return nil, false, nil
 	}
 
-	return nil, false, errors.WithStackTrace(IncludeIsNotABlockErr{parsed: includeBlock})
+	return nil, false, errors.New(IncludeIsNotABlockErr{parsed: includeBlock})
 }
 
-// updateBareIncludeInParsedJSON replaces the include attribute into a block with the label "" in the json. Note that we
+// updateSingleBareIncludeInParsedJSON replaces the include attribute into a block with the label "" in the json. Note that we
 // can directly assign to the map with the single "" key without worrying about the possibility of other include blocks
 // since we will only call this function if there is only one include block, and that is a bare block with no labels.
-func updateSingleBareIncludeInParsedJSON(parsed map[string]interface{}, newVal interface{}) ([]byte, bool, error) {
-	parsed["include"] = map[string]interface{}{bareIncludeKey: newVal}
+func updateSingleBareIncludeInParsedJSON(parsed map[string]any, newVal any) ([]byte, bool, error) {
+	parsed[MetadataInclude] = map[string]any{bareIncludeKey: newVal}
 	updatedBytes, err := json.Marshal(parsed)
-	return updatedBytes, true, errors.WithStackTrace(err)
+
+	return updatedBytes, true, errors.New(err)
 }
 
 // jsonIsIncludeBlock checks if the arbitrary json data is the include block. The data is determined to be an include
@@ -837,8 +957,8 @@ func updateSingleBareIncludeInParsedJSON(parsed map[string]interface{}, newVal i
 // - It is an object
 // - Has the 'path' attribute
 // - The 'path' attribute is a string
-func jsonIsIncludeBlock(jsonData interface{}) bool {
-	typed, isMap := jsonData.(map[string]interface{})
+func jsonIsIncludeBlock(jsonData any) bool {
+	typed, isMap := jsonData.(map[string]any)
 	if isMap {
 		pathAttr, hasPath := typed["path"]
 		if hasPath {
@@ -846,19 +966,39 @@ func jsonIsIncludeBlock(jsonData interface{}) bool {
 			return pathIsString
 		}
 	}
+
 	return false
 }
 
-// copyFieldsMetadata Copy fields metadata between TerragruntConfig instances.
-func copyFieldsMetadata(sourceConfig *TerragruntConfig, targetConfig *TerragruntConfig) {
+// CopyFieldsMetadata Copy fields metadata between TerragruntConfig instances.
+func CopyFieldsMetadata(sourceConfig *TerragruntConfig, targetConfig *TerragruntConfig) {
+	fieldsCopyLocks.Lock(targetConfig.DownloadDir)
+	defer fieldsCopyLocks.Unlock(targetConfig.DownloadDir)
+
 	if sourceConfig.FieldsMetadata != nil {
 		if targetConfig.FieldsMetadata == nil {
-			targetConfig.FieldsMetadata = map[string]map[string]interface{}{}
+			targetConfig.FieldsMetadata = map[string]map[string]any{}
 		}
-		for k, v := range sourceConfig.FieldsMetadata {
-			targetConfig.FieldsMetadata[k] = v
+
+		maps.Copy(targetConfig.FieldsMetadata, sourceConfig.FieldsMetadata)
+	}
+}
+
+// validateGenerateConfigs Validate if exists duplicate generate configs.
+func validateGenerateConfigs(sourceConfig *map[string]codegen.GenerateConfig, targetConfig *map[string]codegen.GenerateConfig) error {
+	var duplicatedNames []string
+
+	for key := range *targetConfig {
+		if _, found := (*sourceConfig)[key]; found {
+			duplicatedNames = append(duplicatedNames, key)
 		}
 	}
+
+	if len(duplicatedNames) != 0 {
+		return DuplicatedGenerateBlocksError{duplicatedNames}
+	}
+
+	return nil
 }
 
 // Custom error types
@@ -870,7 +1010,7 @@ func (err MultipleBareIncludeBlocksErr) Error() string {
 }
 
 type IncludeIsNotABlockErr struct {
-	parsed interface{}
+	parsed any
 }
 
 func (err IncludeIsNotABlockErr) Error() string {
