@@ -5,12 +5,13 @@ import (
 	"strconv"
 )
 
-// Return the kind of the type or Invalid if value is nil
-func KindOf(value interface{}) reflect.Kind {
+// KindOf returns the kind of the type or Invalid if value is nil.
+func KindOf(value any) reflect.Kind {
 	valueType := reflect.TypeOf(value)
 	if valueType == nil {
 		return reflect.Invalid
 	}
+
 	return valueType.Kind()
 }
 
@@ -32,18 +33,21 @@ func KindOf(value interface{}) reflect.Kind {
 //	path ["a", "b", "c"] will return "foo"
 //	path ["a", "d", "1"] will return 2
 //	path ["a", "foo"] will return nil
-func MustWalkTerraformOutput(value interface{}, path ...string) interface{} {
+func MustWalkTerraformOutput(value any, path ...string) any {
 	if value == nil {
 		return nil
 	}
+
 	found := value
 	for _, p := range path {
 		v := reflect.ValueOf(found)
-		switch reflect.TypeOf(found).Kind() {
+
+		switch reflect.TypeOf(found).Kind() { //nolint:exhaustive
 		case reflect.Map:
 			if !v.MapIndex(reflect.ValueOf(p)).IsValid() {
 				return nil
 			}
+
 			found = v.MapIndex(reflect.ValueOf(p)).Interface()
 
 		case reflect.Slice, reflect.Array:
@@ -51,14 +55,17 @@ func MustWalkTerraformOutput(value interface{}, path ...string) interface{} {
 			if err != nil {
 				return nil
 			}
+
 			if v.Len()-1 < i {
 				return nil
 			}
+
 			found = v.Index(i).Interface()
 
 		default:
 			return found
 		}
 	}
+
 	return found
 }
