@@ -1,49 +1,24 @@
-package config
+package config_test
 
 import (
 	"testing"
 
+	"github.com/gruntwork-io/terragrunt/config"
+	"github.com/gruntwork-io/terragrunt/internal/cache"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestStringCacheCreation(t *testing.T) {
-	t.Parallel()
-
-	cache := NewStringCache()
-
-	assert.NotNil(t, cache.Mutex)
-	assert.NotNil(t, cache.Cache)
-
-	assert.Equal(t, 0, len(cache.Cache))
-}
-
-func TestStringCacheOperation(t *testing.T) {
-	t.Parallel()
-
-	cache := NewStringCache()
-
-	value, found := cache.Get("potato")
-
-	assert.False(t, found)
-	assert.Empty(t, value)
-
-	cache.Put("potato", "carrot")
-	value, found = cache.Get("potato")
-
-	assert.True(t, found)
-	assert.NotEmpty(t, value)
-	assert.Equal(t, "carrot", value)
-}
+const testCacheName = "TerragruntConfig"
 
 func TestTerragruntConfigCacheCreation(t *testing.T) {
 	t.Parallel()
 
-	cache := NewTerragruntConfigCache()
+	cache := cache.NewCache[config.TerragruntConfig](testCacheName)
 
 	assert.NotNil(t, cache.Mutex)
 	assert.NotNil(t, cache.Cache)
 
-	assert.Equal(t, 0, len(cache.Cache))
+	assert.Empty(t, cache.Cache)
 }
 
 func TestTerragruntConfigCacheOperation(t *testing.T) {
@@ -51,19 +26,20 @@ func TestTerragruntConfigCacheOperation(t *testing.T) {
 
 	testCacheKey := "super-safe-cache-key"
 
-	cache := NewTerragruntConfigCache()
+	ctx := t.Context()
+	cache := cache.NewCache[config.TerragruntConfig](testCacheName)
 
-	actualResult, found := cache.Get(testCacheKey)
+	actualResult, found := cache.Get(ctx, testCacheKey)
 
 	assert.False(t, found)
 	assert.Empty(t, actualResult)
 
-	stubTerragruntConfig := TerragruntConfig{
+	stubTerragruntConfig := config.TerragruntConfig{
 		IsPartial: true, // Any random property will be sufficient
 	}
 
-	cache.Put(testCacheKey, stubTerragruntConfig)
-	actualResult, found = cache.Get(testCacheKey)
+	cache.Put(ctx, testCacheKey, stubTerragruntConfig)
+	actualResult, found = cache.Get(ctx, testCacheKey)
 
 	assert.True(t, found)
 	assert.NotEmpty(t, actualResult)
