@@ -1,80 +1,80 @@
-package config
+package config_test
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/hashicorp/hcl/v2/hclparse"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/zclconf/go-cty/cty/gocty"
 
-	"github.com/gruntwork-io/terragrunt/errors"
+	"github.com/gruntwork-io/terragrunt/config"
+	"github.com/gruntwork-io/terragrunt/config/hclparse"
+	"github.com/gruntwork-io/terragrunt/internal/errors"
+	"github.com/gruntwork-io/terragrunt/test/helpers/logger"
 )
 
 func TestEvaluateLocalsBlock(t *testing.T) {
 	t.Parallel()
 
 	terragruntOptions := mockOptionsForTest(t)
-	mockFilename := "terragrunt.hcl"
 
-	parser := hclparse.NewParser()
-	file, err := parseHcl(parser, LocalsTestConfig, mockFilename)
+	file, err := hclparse.NewParser().ParseFromString(LocalsTestConfig, config.DefaultTerragruntConfigPath)
 	require.NoError(t, err)
 
-	evaluatedLocals, err := evaluateLocalsBlock(terragruntOptions, parser, file, mockFilename, nil, nil)
+	ctx := config.NewParsingContext(t.Context(), logger.CreateLogger(), terragruntOptions)
+	evaluatedLocals, err := config.EvaluateLocalsBlock(ctx, logger.CreateLogger(), file)
 	require.NoError(t, err)
 
 	var actualRegion string
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["region"], &actualRegion))
-	assert.Equal(t, actualRegion, "us-east-1")
+	assert.Equal(t, "us-east-1", actualRegion)
 
 	var actualS3Url string
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["s3_url"], &actualS3Url))
-	assert.Equal(t, actualS3Url, "com.amazonaws.us-east-1.s3")
+	assert.Equal(t, "com.amazonaws.us-east-1.s3", actualS3Url)
 
 	var actualX float64
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["x"], &actualX))
-	assert.Equal(t, actualX, float64(1))
+	assert.InEpsilon(t, float64(1), actualX, 0.0000001)
 
-	var actualY float64
-	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["y"], &actualY))
-	assert.Equal(t, actualY, float64(2))
+	var actualY float64                                                    //codespell:ignore
+	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["y"], &actualY)) //codespell:ignore
+	assert.InEpsilon(t, float64(2), actualY, 0.0000001)                    //codespell:ignore
 
 	var actualZ float64
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["z"], &actualZ))
-	assert.Equal(t, actualZ, float64(3))
+	assert.InEpsilon(t, float64(3), actualZ, 0.0000001)
 
 	var actualFoo struct{ First Foo }
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["foo"], &actualFoo))
-	assert.Equal(t, actualFoo.First, Foo{
+	assert.Equal(t, Foo{
 		Region: "us-east-1",
 		Foo:    "bar",
-	})
+	}, actualFoo.First)
 
 	var actualBar string
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["bar"], &actualBar))
-	assert.Equal(t, actualBar, "us-east-1")
+	assert.Equal(t, "us-east-1", actualBar)
 }
 
 func TestEvaluateLocalsBlockMultiDeepReference(t *testing.T) {
 	t.Parallel()
 
 	terragruntOptions := mockOptionsForTest(t)
-	mockFilename := "terragrunt.hcl"
 
-	parser := hclparse.NewParser()
-	file, err := parseHcl(parser, LocalsTestMultiDeepReferenceConfig, mockFilename)
+	file, err := hclparse.NewParser().ParseFromString(LocalsTestMultiDeepReferenceConfig, config.DefaultTerragruntConfigPath)
 	require.NoError(t, err)
 
-	evaluatedLocals, err := evaluateLocalsBlock(terragruntOptions, parser, file, mockFilename, nil, nil)
+	ctx := config.NewParsingContext(t.Context(), logger.CreateLogger(), terragruntOptions)
+	evaluatedLocals, err := config.EvaluateLocalsBlock(ctx, logger.CreateLogger(), file)
 	require.NoError(t, err)
 
 	expected := "a"
 
 	var actualA string
 	require.NoError(t, gocty.FromCtyValue(evaluatedLocals["a"], &actualA))
-	assert.Equal(t, actualA, expected)
+	assert.Equal(t, expected, actualA)
 
 	testCases := []string{
 		"b",
@@ -87,12 +87,12 @@ func TestEvaluateLocalsBlockMultiDeepReference(t *testing.T) {
 		"i",
 		"j",
 	}
-	for _, testCase := range testCases {
-		expected = fmt.Sprintf("%s/%s", expected, testCase)
+	for _, tc := range testCases {
+		expected = fmt.Sprintf("%s/%s", expected, tc)
 
 		var actual string
-		require.NoError(t, gocty.FromCtyValue(evaluatedLocals[testCase], &actual))
-		assert.Equal(t, actual, expected)
+		require.NoError(t, gocty.FromCtyValue(evaluatedLocals[tc], &actual))
+		assert.Equal(t, expected, actual)
 	}
 }
 
@@ -100,17 +100,16 @@ func TestEvaluateLocalsBlockImpossibleWillFail(t *testing.T) {
 	t.Parallel()
 
 	terragruntOptions := mockOptionsForTest(t)
-	mockFilename := "terragrunt.hcl"
 
-	parser := hclparse.NewParser()
-	file, err := parseHcl(parser, LocalsTestImpossibleConfig, mockFilename)
+	file, err := hclparse.NewParser().ParseFromString(LocalsTestImpossibleConfig, config.DefaultTerragruntConfigPath)
 	require.NoError(t, err)
 
-	_, err = evaluateLocalsBlock(terragruntOptions, parser, file, mockFilename, nil, nil)
+	ctx := config.NewParsingContext(t.Context(), logger.CreateLogger(), terragruntOptions)
+	_, err = config.EvaluateLocalsBlock(ctx, logger.CreateLogger(), file)
 	require.Error(t, err)
 
-	switch errors.Unwrap(err).(type) {
-	case CouldNotEvaluateAllLocalsError:
+	switch errors.Unwrap(err).(type) { //nolint:errorlint
+	case config.CouldNotEvaluateAllLocalsError:
 	default:
 		t.Fatalf("Did not get expected error: %s", err)
 	}
@@ -120,13 +119,12 @@ func TestEvaluateLocalsBlockMultipleLocalsBlocksWillFail(t *testing.T) {
 	t.Parallel()
 
 	terragruntOptions := mockOptionsForTest(t)
-	mockFilename := "terragrunt.hcl"
 
-	parser := hclparse.NewParser()
-	file, err := parseHcl(parser, MultipleLocalsBlockConfig, mockFilename)
+	file, err := hclparse.NewParser().ParseFromString(MultipleLocalsBlockConfig, config.DefaultTerragruntConfigPath)
 	require.NoError(t, err)
 
-	_, err = evaluateLocalsBlock(terragruntOptions, parser, file, mockFilename, nil, nil)
+	ctx := config.NewParsingContext(t.Context(), logger.CreateLogger(), terragruntOptions)
+	_, err = config.EvaluateLocalsBlock(ctx, logger.CreateLogger(), file)
 	require.Error(t, err)
 }
 
