@@ -1,18 +1,18 @@
-package test
+package test_test
 
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"testing"
 
+	"github.com/gruntwork-io/terragrunt/test/helpers"
 	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 const (
-	registryFixturePath                          = "fixture-tfr"
+	registryFixturePath                          = "fixtures/tfr"
 	registryFixtureRootModulePath                = "root"
 	registryFixtureRootShorthandModulePath       = "root-shorthand"
 	registryFixtureSubdirModulePath              = "subdir"
@@ -24,7 +24,7 @@ func TestTerraformRegistryFetchingRootModule(t *testing.T) {
 	testTerraformRegistryFetching(t, registryFixtureRootModulePath, "root_null_resource")
 }
 
-func TestTerraformRegistryFetchingRootShorthandModule(t *testing.T) {
+func TestRegistryFetchingRootShorthandModule(t *testing.T) {
 	t.Parallel()
 	testTerraformRegistryFetching(t, registryFixtureRootShorthandModulePath, "root_null_resource")
 }
@@ -40,18 +40,19 @@ func TestTerraformRegistryFetchingSubdirWithReferenceModule(t *testing.T) {
 }
 
 func testTerraformRegistryFetching(t *testing.T, modPath, expectedOutputKey string) {
+	t.Helper()
+
 	modFullPath := util.JoinPath(registryFixturePath, modPath)
-	cleanupTerraformFolder(t, modFullPath)
-	runTerragrunt(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-log-level debug --terragrunt-working-dir %s", modFullPath))
+	helpers.CleanupTerraformFolder(t, modFullPath)
+	helpers.RunTerragrunt(t, "terragrunt apply -auto-approve --non-interactive --log-level trace --working-dir "+modFullPath)
 
 	stdout := bytes.Buffer{}
 	stderr := bytes.Buffer{}
-	err := runTerragruntCommand(t, fmt.Sprintf("terragrunt output -no-color -json --terragrunt-non-interactive --terragrunt-log-level debug --terragrunt-working-dir %s", modFullPath), &stdout, &stderr)
+	err := helpers.RunTerragruntCommand(t, "terragrunt output -no-color -json --non-interactive --log-level trace --working-dir "+modFullPath, &stdout, &stderr)
 	require.NoError(t, err)
 
-	outputs := map[string]TerraformOutput{}
-	require.NoError(t, json.Unmarshal([]byte(stdout.String()), &outputs))
+	outputs := map[string]helpers.TerraformOutput{}
+	require.NoError(t, json.Unmarshal(stdout.Bytes(), &outputs))
 	_, hasOutput := outputs[expectedOutputKey]
 	assert.True(t, hasOutput)
-
 }
