@@ -1,8 +1,10 @@
-package util
+package util_test
 
 import (
+	"strconv"
 	"testing"
 
+	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -16,24 +18,28 @@ func TestMatchesAny(t *testing.T) {
 	}
 
 	testCases := []struct {
-		list     []string
 		element  string
+		list     []string
 		expected bool
 	}{
-		{nil, "", false},
-		{[]string{}, "", false},
-		{[]string{}, "foo", false},
-		{[]string{"foo"}, "kafoot", true},
-		{[]string{"bar", "foo", ".*Failed to load backend.*TLS handshake timeout.*"}, "Failed to load backend: Error...:...  TLS handshake timeout", true},
-		{[]string{"bar", "foo", ".*Failed to load backend.*TLS handshake timeout.*"}, "Failed to load backend: Error...:...  TLxS handshake timeout", false},
-		{[]string{"(?s).*Failed to load state.*dial tcp.*timeout.*"}, realWorldErrorMessages[0], true},
-		{[]string{"(?s).*Creating metric alarm failed.*request to update this alarm is in progress.*"}, realWorldErrorMessages[1], true},
-		{[]string{"(?s).*Error configuring the backend.*TLS handshake timeout.*"}, realWorldErrorMessages[2], true},
+		{list: nil, element: "", expected: false},
+		{list: []string{}, element: "", expected: false},
+		{list: []string{}, element: "foo", expected: false},
+		{list: []string{"foo"}, element: "kafoot", expected: true},
+		{list: []string{"bar", "foo", ".*Failed to load backend.*TLS handshake timeout.*"}, element: "Failed to load backend: Error...:...  TLS handshake timeout", expected: true},
+		{list: []string{"bar", "foo", ".*Failed to load backend.*TLS handshake timeout.*"}, element: "Failed to load backend: Error...:...  TLxS handshake timeout", expected: false},
+		{list: []string{"(?s).*Failed to load state.*dial tcp.*timeout.*"}, element: realWorldErrorMessages[0], expected: true},
+		{list: []string{"(?s).*Creating metric alarm failed.*request to update this alarm is in progress.*"}, element: realWorldErrorMessages[1], expected: true},
+		{list: []string{"(?s).*Error configuring the backend.*TLS handshake timeout.*"}, element: realWorldErrorMessages[2], expected: true},
 	}
 
-	for _, testCase := range testCases {
-		actual := MatchesAny(testCase.list, testCase.element)
-		assert.Equal(t, testCase.expected, actual, "For list %v and element %s", testCase.list, testCase.element)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.MatchesAny(tc.list, tc.element)
+			assert.Equal(t, tc.expected, actual, "For list %v and element %s", tc.list, tc.element)
+		})
 	}
 }
 
@@ -41,21 +47,25 @@ func TestListContainsElement(t *testing.T) {
 	t.Parallel()
 
 	testCases := []struct {
-		list     []string
 		element  string
+		list     []string
 		expected bool
 	}{
-		{[]string{}, "", false},
-		{[]string{}, "foo", false},
-		{[]string{"foo"}, "foo", true},
-		{[]string{"bar", "foo", "baz"}, "foo", true},
-		{[]string{"bar", "foo", "baz"}, "nope", false},
-		{[]string{"bar", "foo", "baz"}, "", false},
+		{list: []string{}, element: "", expected: false},
+		{list: []string{}, element: "foo", expected: false},
+		{list: []string{"foo"}, element: "foo", expected: true},
+		{list: []string{"bar", "foo", "baz"}, element: "foo", expected: true},
+		{list: []string{"bar", "foo", "baz"}, element: "nope", expected: false},
+		{list: []string{"bar", "foo", "baz"}, element: "", expected: false},
 	}
 
-	for _, testCase := range testCases {
-		actual := ListContainsElement(testCase.list, testCase.element)
-		assert.Equal(t, testCase.expected, actual, "For list %v and element %s", testCase.list, testCase.element)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.ListContainsElement(tc.list, tc.element)
+			assert.Equal(t, tc.expected, actual, "For list %v and element %s", tc.list, tc.element)
+		})
 	}
 }
 
@@ -77,9 +87,13 @@ func TestListEquals(t *testing.T) {
 		{[]string{""}, []string{""}, true},
 		{[]string{"foo", "bar"}, []string{"foo", "bar"}, true},
 	}
-	for _, testCase := range testCases {
-		actual := ListEquals(testCase.a, testCase.b)
-		assert.Equal(t, testCase.expected, actual, "For list %v and list %v", testCase.a, testCase.b)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.ListEquals(tc.a, tc.b)
+			assert.Equal(t, tc.expected, actual, "For list %v and list %v", tc.a, tc.b)
+		})
 	}
 }
 
@@ -114,9 +128,13 @@ func TestListContainsSublist(t *testing.T) {
 		{[]string{"zim", "gee", "foo", "bar"}, []string{"gee", "foo", "bar"}, true},
 	}
 
-	for _, testCase := range testCases {
-		actual := ListContainsSublist(testCase.list, testCase.sublist)
-		assert.Equal(t, testCase.expected, actual, "For list %v and sublist %v", testCase.list, testCase.sublist)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.ListContainsSublist(tc.list, tc.sublist)
+			assert.Equal(t, tc.expected, actual, "For list %v and sublist %v", tc.list, tc.sublist)
+		})
 	}
 }
 
@@ -141,9 +159,13 @@ func TestListHasPrefix(t *testing.T) {
 		{[]string{"foo", "bar"}, []string{"foo", "bar"}, true},
 		{[]string{"foo", "bar", "biz"}, []string{"foo", "bar"}, true},
 	}
-	for _, testCase := range testCases {
-		actual := ListHasPrefix(testCase.list, testCase.prefix)
-		assert.Equal(t, testCase.expected, actual, "For list %v and prefix %v", testCase.list, testCase.prefix)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.ListHasPrefix(tc.list, tc.prefix)
+			assert.Equal(t, tc.expected, actual, "For list %v and prefix %v", tc.list, tc.prefix)
+		})
 	}
 }
 
@@ -155,18 +177,22 @@ func TestRemoveElementFromList(t *testing.T) {
 		element  string
 		expected []string
 	}{
-		{[]string{}, "", []string{}},
-		{[]string{}, "foo", []string{}},
-		{[]string{"foo"}, "foo", []string{}},
+		{[]string{}, "", nil},
+		{[]string{}, "foo", nil},
+		{[]string{"foo"}, "foo", nil},
 		{[]string{"bar"}, "foo", []string{"bar"}},
 		{[]string{"bar", "foo", "baz"}, "foo", []string{"bar", "baz"}},
 		{[]string{"bar", "foo", "baz"}, "nope", []string{"bar", "foo", "baz"}},
 		{[]string{"bar", "foo", "baz"}, "", []string{"bar", "foo", "baz"}},
 	}
 
-	for _, testCase := range testCases {
-		actual := RemoveElementFromList(testCase.list, testCase.element)
-		assert.Equal(t, testCase.expected, actual, "For list %v and element %s", testCase.list, testCase.element)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			actual := util.RemoveElementFromList(tc.list, tc.element)
+			assert.Equal(t, tc.expected, actual, "For list %v and element %s", tc.list, tc.element)
+		})
 	}
 }
 
@@ -187,13 +213,18 @@ func TestRemoveDuplicatesFromList(t *testing.T) {
 		{[]string{"foo", "bar", "foobar", "foo", "bar"}, []string{"foobar", "foo", "bar"}, true},
 	}
 
-	for _, testCase := range testCases {
-		f := RemoveDuplicatesFromList
-		if testCase.reverse {
-			f = RemoveDuplicatesFromListKeepLast
-		}
-		assert.Equal(t, f(testCase.list), testCase.expected, "For list %v", testCase.list)
-		t.Logf("%v passed", testCase.list)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			f := util.RemoveDuplicatesFromList[[]string]
+			if tc.reverse {
+				f = util.RemoveDuplicatesFromListKeepLast[[]string]
+			}
+
+			assert.Equal(t, tc.expected, f(tc.list), "For list %v", tc.list)
+			t.Logf("%v passed", tc.list)
+		})
 	}
 }
 
@@ -201,17 +232,21 @@ func TestCommaSeparatedStrings(t *testing.T) {
 	t.Parallel()
 
 	testCases := []struct {
-		list     []string
 		expected string
+		list     []string
 	}{
-		{[]string{}, ``},
-		{[]string{"foo"}, `"foo"`},
-		{[]string{"foo", "bar"}, `"foo", "bar"`},
+		{list: []string{}, expected: ``},
+		{list: []string{"foo"}, expected: `"foo"`},
+		{list: []string{"foo", "bar"}, expected: `"foo", "bar"`},
 	}
 
-	for _, testCase := range testCases {
-		assert.Equal(t, CommaSeparatedStrings(testCase.list), testCase.expected, "For list %v", testCase.list)
-		t.Logf("%v passed", testCase.list)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			assert.Equal(t, tc.expected, util.CommaSeparatedStrings(tc.list), "For list %v", tc.list)
+			t.Logf("%v passed", tc.list)
+		})
 	}
 }
 
@@ -219,70 +254,50 @@ func TestStringListInsert(t *testing.T) {
 	t.Parallel()
 
 	testCases := []struct {
-		list     []string
 		element  string
-		index    int
+		list     []string
 		expected []string
+		index    int
 	}{
-		{[]string{}, "foo", 0, []string{"foo"}},
-		{[]string{"a", "c", "d"}, "b", 1, []string{"a", "b", "c", "d"}},
-		{[]string{"b", "c", "d"}, "a", 0, []string{"a", "b", "c", "d"}},
-		{[]string{"a", "b", "d"}, "c", 2, []string{"a", "b", "c", "d"}},
+		{list: []string{}, element: "foo", index: 0, expected: []string{"foo"}},
+		{list: []string{"a", "c", "d"}, element: "b", index: 1, expected: []string{"a", "b", "c", "d"}},
+		{list: []string{"b", "c", "d"}, element: "a", index: 0, expected: []string{"a", "b", "c", "d"}},
+		{list: []string{"a", "b", "d"}, element: "c", index: 2, expected: []string{"a", "b", "c", "d"}},
 	}
 
-	for _, testCase := range testCases {
-		assert.Equal(t, testCase.expected, StringListInsert(testCase.list, testCase.element, testCase.index), "For list %v", testCase.list)
-		t.Logf("%v passed", testCase.list)
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			assert.Equal(t, tc.expected, util.StringListInsert(tc.list, tc.element, tc.index), "For list %v", tc.list)
+			t.Logf("%v passed", tc.list)
+		})
 	}
 }
 
-func TestKeyValuePairStringListToMap(t *testing.T) {
+func TestMapToSlice(t *testing.T) {
 	t.Parallel()
 
-	testCases := []struct {
-		name     string
-		input    []string
-		splitter func(s, sep string) []string
-		output   map[string]string
-	}{
-		{
-			"base",
-			[]string{"foo=bar", "baz=carol"},
-			SplitUrls,
-			map[string]string{
-				"foo": "bar",
-				"baz": "carol",
-			},
-		},
-		{
-			"special_chars",
-			[]string{"ssh://git@github.com=/path/to/local"},
-			SplitUrls,
-			map[string]string{"ssh://git@github.com": "/path/to/local"},
-		},
-		{
-			"with_tags",
-			[]string{"ssh://git@github.com=ssh://git@github.com/test.git?ref=feature"},
-			SplitUrls,
-			map[string]string{"ssh://git@github.com": "ssh://git@github.com/test.git?ref=feature"},
-		},
-		{
-			"empty",
-			[]string{},
-			SplitUrls,
-			map[string]string{},
-		},
-	}
+	t.Run("Empty Map", func(t *testing.T) {
+		t.Parallel()
 
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			actualOutput, err := KeyValuePairStringListToMap(testCase.input, testCase.splitter)
-			assert.NoError(t, err)
-			assert.Equal(
-				t,
-				testCase.output,
-				actualOutput,
-			)
-		})
-	}
+		m := make(map[string]*int)
+
+		result := util.MapToSlice(m)
+		if len(result) != 0 {
+			t.Errorf("Expected empty slice, got %v", result)
+		}
+	})
+
+	t.Run("Single Element Map", func(t *testing.T) {
+		t.Parallel()
+
+		val := 42
+		m := map[string]*int{"key1": &val}
+
+		result := util.MapToSlice(m)
+		if len(result) != 1 || result[0] != &val {
+			t.Errorf("Expected slice with one element %v, got %v", &val, result)
+		}
+	})
 }
