@@ -1,11 +1,13 @@
-package util
+package util_test
 
 import (
 	"fmt"
 	"testing"
 	"time"
 
+	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestParseTimestamp(t *testing.T) {
@@ -22,20 +24,18 @@ func TestParseTimestamp(t *testing.T) {
 		{"2017-11-22 00:00:00Z", time.Time{}, `not a valid RFC3339 timestamp: missing required time introducer 'T'`},
 	}
 
-	for _, testCase := range testCases {
-		testCase := testCase
-
-		t.Run(fmt.Sprintf("ParseTimestamp(%#v)", testCase.arg), func(t *testing.T) {
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("ParseTimestamp(%#v)", tc.arg), func(t *testing.T) {
 			t.Parallel()
 
-			actual, err := ParseTimestamp(testCase.arg)
-			if testCase.err != "" {
-				assert.EqualError(t, err, testCase.err)
+			actual, err := util.ParseTimestamp(tc.arg)
+			if tc.err != "" {
+				require.EqualError(t, err, tc.err)
 			} else {
-				assert.NoError(t, err)
+				require.NoError(t, err)
 			}
 
-			assert.Equal(t, testCase.value, actual)
+			assert.Equal(t, tc.value, actual)
 		})
 	}
 }
