@@ -1,11 +1,14 @@
-package util
+package util_test
 
 import (
 	"bytes"
 	"errors"
+	"strconv"
 	"testing"
 
+	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestPrefixWriter(t *testing.T) {
@@ -13,41 +16,101 @@ func TestPrefixWriter(t *testing.T) {
 
 	testCases := []struct {
 		prefix   string
-		values   []string
 		expected string
+		values   []string
 	}{
-		{"p1 ", []string{"a", "b"}, "p1 ab"},
-		{"p2 ", []string{"a", "b"}, "p2 ab"},
-
-		{"", []string{"a", "b"}, "ab"},
-
-		{"p1 ", []string{"a", "b\n"}, "p1 ab\n"},
-		{"p1 ", []string{"a\n", "b"}, "p1 a\np1 b"},
-		{"p1 ", []string{"a\n", "b\n"}, "p1 a\np1 b\n"},
-		{"p1 ", []string{"a", "b", "c", "def"}, "p1 abcdef"},
-		{"p1 ", []string{"a", "b\n", "c", "def"}, "p1 ab\np1 cdef"},
-		{"p1 ", []string{"a", "b\nc", "def"}, "p1 ab\np1 cdef"},
-		{"p1 ", []string{"ab", "cd", "ef", "gh\n"}, "p1 abcdefgh\n"},
-		{"p1 ", []string{"ab", "cd\n", "ef", "gh\n"}, "p1 abcd\np1 efgh\n"},
-		{"p1 ", []string{"ab", "cd", "e\nf", "gh\n"}, "p1 abcde\np1 fgh\n"},
-		{"p1 ", []string{"ab", "cd", "ef\n", "gh\n"}, "p1 abcdef\np1 gh\n"},
-		{"p1 ", []string{"ab\ncd\nef\ngh\n"}, "p1 ab\np1 cd\np1 ef\np1 gh\n"},
-		{"p1 ", []string{"ab\n\n\ngh\n"}, "p1 ab\np1 \np1 \np1 gh\n"},
-
-		{"p1 ", []string{""}, ""},
-		{"p1 ", []string{"\n"}, "p1 \n"},
-		{"p1\n", []string{"\n"}, "p1\n\n"},
+		{
+			prefix:   "p1 ",
+			expected: "p1 ab",
+			values:   []string{"a", "b"},
+		},
+		{
+			prefix:   "p2 ",
+			expected: "p2 ab",
+			values:   []string{"a", "b"},
+		},
+		{
+			prefix:   "",
+			expected: "ab",
+			values:   []string{"a", "b"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 ab\n",
+			values:   []string{"a", "b\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 a\np1 b",
+			values:   []string{"a\n", "b"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 a\np1 b\n",
+			values:   []string{"a\n", "b\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 abcdef",
+			values:   []string{"a", "b", "c", "def"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 abcdefgh\n",
+			values:   []string{"ab", "cd", "ef", "gh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 abcd\np1 efgh\n",
+			values:   []string{"ab", "cd\n", "ef", "gh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 abcde\np1 fgh\n",
+			values:   []string{"ab", "cd", "e\nf", "gh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 abcdef\np1 gh\n",
+			values:   []string{"ab", "cd", "ef\n", "gh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 ab\np1 cd\np1 ef\np1 gh\n",
+			values:   []string{"ab\ncd\nef\ngh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 ab\np1 \np1 \np1 gh\n",
+			values:   []string{"ab\n\n\ngh\n"},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "",
+			values:   []string{""},
+		},
+		{
+			prefix:   "p1 ",
+			expected: "p1 \n",
+			values:   []string{"\n"},
+		},
 	}
 
-	for _, testCase := range testCases {
-		var b bytes.Buffer
-		pw := PrefixedWriter(&b, testCase.prefix)
-		for _, input := range testCase.values {
-			written, err := pw.Write([]byte(input))
-			assert.NoError(t, err)
-			assert.Equal(t, written, len(input))
-		}
-		assert.Equal(t, testCase.expected, b.String())
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			var b bytes.Buffer
+
+			pw := util.PrefixedWriter(&b, tc.prefix)
+			for _, input := range tc.values {
+				written, err := pw.Write([]byte(input))
+				require.NoError(t, err)
+				assert.Len(t, input, written)
+			}
+
+			assert.Equal(t, tc.expected, b.String())
+		})
 	}
 }
 
@@ -62,18 +125,26 @@ func TestPrefixWriterFail(t *testing.T) {
 
 	testCases := []struct {
 		prefix   string
-		values   []string
 		expected string
+		values   []string
 	}{
-		{"p1 ", []string{"a", "b"}, "p1 ab"},
+		{
+			prefix:   "p1 ",
+			expected: "p1 ab",
+			values:   []string{"a", "b"},
+		},
 	}
 
-	for _, testCase := range testCases {
-		pw := PrefixedWriter(&FailingWriter{}, testCase.prefix)
-		for _, input := range testCase.values {
-			written, err := pw.Write([]byte(input))
-			assert.Error(t, err)
-			assert.Equal(t, written, 0)
-		}
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Parallel()
+
+			pw := util.PrefixedWriter(&FailingWriter{}, tc.prefix)
+			for _, input := range tc.values {
+				written, err := pw.Write([]byte(input))
+				require.Error(t, err)
+				assert.Empty(t, written)
+			}
+		})
 	}
 }
